requirements.txt
absl-py==0.6.1
astor==0.7.1
beautifulsoup4==4.6.3
certifi==2018.11.29
chardet==3.0.4
cycler==0.10.0
gast==0.2.0
grpcio==1.17.1
h5py==2.8.0
idna==2.8
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==3.0.2
numpy==1.15.4
opencv-python==3.4.4.19
pandas==0.23.4
Pillow==5.3.0
protobuf==3.6.1
pyparsing==2.3.0
PyQt5==5.11.3
PyQt5-sip==4.19.13
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.13
requests==2.21.0
scipy==1.1.0
seaborn==0.9.0
six==1.12.0
tensorboard==1.12.1
tensorflow==1.12.0
termcolor==1.1.0
urllib3==1.24.1
Werkzeug==0.14.1
wikipedia==1.4.0
